#24 offerId-myOfferId switch clarified
For 7(messaging) there is a trick performed by the API: it swaps myOfferId with offerId.
It is worth to notice the fact that chat performed between offers, not between users, so if a user sends a message to an offer he specifies his own offer as it is seen on the recipient's side. So outgoing offer id (my offer) becomes incoming offer id for my party in his/her list and vice versa - my party sees my offer as an item in incoming offers list, so my offer id is just another offer for my party.